feat(callout-admin-ux): Moved intro message to sidebar tabs
Introduce an intro message tab in the sidebar for callouts. 
This change includes the addition of a new `IntroMessageTab` 
component and updates to the sidebar tabs interface to 
accommodate the new tab. The intro text is now editable 
via a rich text editor, enhancing the user experience 
by allowing customization of introductory content.