feat(vue/AppToggleField): add state-specific descriptions
This change adds support for separate descriptions for the enabled and disabled states of the `AppToggleField` component. Previously, the component only had a single description that was displayed regardless of the toggle state. The new `enabledDescription` and `disabledDescription` props allow for more contextual information to be provided to the user.

The commit also updates the stories to demonstrate the new functionality and provides examples of how the state-specific descriptions can be used in different use cases.