refactor(address): Remove intermediary address types
This commit removes all intermediary address types in favour of
CalloutResponseAddressAnswer. This type represents the underyling
data structure in the database, we should be converting all other
geocoding data formats into this.

This commit introduces some backwards incompatible changes to the
CalloutResponseAddressAnswer type which will need to be addressed with a
database migration.